server/requirements.txt
flask
flask-cors
python-dotenv
pymupdf
requests
exa-py
